requirements.txt
# requirements.txt
pyspark==4.0.1
psycopg2-binary==2.9.7

pandas==2.1.3
pyarrow==12.0.1
pytest==7.3.2

apache-airflow-providers-apache-spark